[mock] change folder name
