+++ src/main/java/com/example/capstone_project/controller/TermController.java
@@ -3,10 +3,10 @@ package com.example.capstone_project.controller;
 import com.example.capstone_project.controller.body.term.create.CreateTermBody;
 import com.example.capstone_project.controller.body.term.delete.DeleteTermBody;
 import com.example.capstone_project.controller.body.term.update.UpdateTermBody;
-import com.example.capstone_project.controller.responses.term.get_plans.PlanStatusResponse;
-import com.example.capstone_project.controller.responses.term.get_termDetail.TermDetailResponse;
-import com.example.capstone_project.controller.responses.term.get_plans.TermPlanDetailResponse;
-import com.example.capstone_project.controller.responses.term.get_termDetail.TermStatusResponse;
+import com.example.capstone_project.controller.responses.term.getPlans.PlanStatusResponse;
+import com.example.capstone_project.controller.responses.term.getTermDetail.TermDetailResponse;
+import com.example.capstone_project.controller.responses.term.getPlans.TermPlanDetailResponse;
+import com.example.capstone_project.controller.responses.term.getTermDetail.TermStatusResponse;
 import com.example.capstone_project.entity.TermDuration;
 import com.example.capstone_project.service.term.TermService;
 import com.example.capstone_project.utils.enums.TermCode;
@@ -17,8 +17,6 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
